Add RepositoryFileRequest
